refactor: move check-metric helpers from metrics/utils to metrics-prometheus
move check-metrics.service.ts and check-metric-labels.ts to
metrics-prometheus/ and update all imports across deal/, retrieval/,
data-retention/, deal-addons/, and metrics-prometheus.module.ts.
